Don't let navigation overlay consume click event
So it won't take two actions to dismiss overlay & app bar.
