+++ app/src/main/kotlin/com/hippo/ehviewer/ui/reader/NavigationOverlay.kt
@@ -5,20 +5,13 @@ import androidx.compose.animation.core.tween
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -26,34 +19,20 @@ import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
-import eu.kanade.tachiyomi.ui.reader.viewer.ViewerNavigation
+import eu.kanade.tachiyomi.ui.reader.viewer.NavigationRegions
 
 @Composable
 fun NavigationOverlay(
+    visible: Boolean,
     regions: NavigationRegions,
-    showOnStart: Boolean,
-    onDismiss: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    var firstLaunch by remember { mutableStateOf(true) }
-    var visible by remember { mutableStateOf(showOnStart) }
-    LaunchedEffect(regions) {
-        if (firstLaunch) {
-            firstLaunch = false
-        } else {
-            visible = true
-        }
-    }
     AnimatedVisibility(
         visible = visible && regions.isNotEmpty(),
-        modifier = modifier.clickable(interactionSource = null, indication = null) {
-            onDismiss()
-            visible = false
-        }.fillMaxSize(),
         enter = fadeIn(AnimationSpec),
         exit = fadeOut(AnimationSpec),
     ) {
-        BoxWithConstraints {
+        BoxWithConstraints(modifier) {
             val fontSize = MaterialTheme.typography.headlineSmall.fontSize
             val style = TextStyle(
                 color = Color.White,
@@ -82,7 +61,5 @@ fun NavigationOverlay(
         }
     }
 }
-
-typealias NavigationRegions = List<ViewerNavigation.Region>
 
 private val AnimationSpec = tween<Float>(1000)
