+++ app/src/main/kotlin/com/hippo/ehviewer/ui/reader/Utils.kt
@@ -17,9 +17,6 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.layout.FixedScale
 import com.hippo.ehviewer.Settings
 import eu.kanade.tachiyomi.ui.reader.setting.OrientationType
-import eu.kanade.tachiyomi.ui.reader.setting.ReadingModeType
-import eu.kanade.tachiyomi.ui.reader.setting.ReadingModeType.RIGHT_TO_LEFT
-import eu.kanade.tachiyomi.ui.reader.setting.ReadingModeType.VERTICAL
 import kotlinx.coroutines.flow.onCompletion
 
 fun Window.updateKeepScreenOn(enabled: Boolean) {
@@ -66,10 +63,10 @@ fun Activity.setCustomBrightnessValue(value: Int) {
 }
 
 @Stable
-fun Alignment.Companion.fromPreferences(value: Int, mode: ReadingModeType) = when (value) {
-    1 -> when (mode) {
-        VERTICAL -> CenterHorizontally
-        RIGHT_TO_LEFT -> AbsoluteAlignment.Right
+fun Alignment.Companion.fromPreferences(value: Int, isRtl: Boolean, isVertical: Boolean) = when (value) {
+    1 -> when {
+        isVertical -> CenterHorizontally
+        isRtl -> AbsoluteAlignment.Right
         else -> AbsoluteAlignment.Left
     }
     2 -> AbsoluteAlignment.Left
