+++ app/src/main/kotlin/eu/kanade/tachiyomi/ui/reader/viewer/ViewerNavigation.kt
@@ -50,3 +50,8 @@ abstract class ViewerNavigation {
         }
     }
 }
+
+typealias NavigationRegions = List<ViewerNavigation.Region>
+
+fun NavigationRegions.getAction(x: Float, y: Float) =
+    find { it.rectF.contains(x, y) }?.type ?: ViewerNavigation.NavigationRegion.MENU
